Add build and packaging for mygov deployment pipeline
Create a module (platform) for building a 'platform' webapp.
See Bloomreach documentation: "Create Separate Distributions for the
Authoring and Delivery Applications" for more information.

Add modules to build debian packages for the CMS, sites, platform.

Add module (publishing-libs) to build a debian package containing
libraries that need to be installed on the Tomcat shared/common
classpaths.

Configure parent pom with values that would normally be inherited
from an organisation parent pom. Unfortunately, Hippo assumes that
one of their pom files will be used as a parent, so this needs to
be duplicated.

Remove repositories section from POM. The use of this is discouraged
in favour of using a repository manager, in case the URLs change.