MGS-5549 Add 404 page
There are 2 types of 404 pages - one for pages and a requests that
dont match anything in the sitemap.

The second kind is a fallback jsp page.  This is used as a fallback
in case hippo does not otherwise handle the error.  It is also used
for binary requests.  Currently this page is not styled but we will
add some basic styling.