MGS-5651 Remove sitemap entry for site login page
Hippo provides a /login sitemap entry for the site by default. This is
used for delivery tier authorisation - e.g. restricting a mount or
sitemap item to a specific role or group.

We don't currently use this functionality. This commit removes the
ability to access this /login page and a see an 'authentication failed'
message or the following '/j_security_check' page.