MGS-5837 allow content to appear in multiple categories
Adds a new document type called 'mirror' which can be added to
any category. This has a field which can point at another article
or category.

When a mirror is added to a category it causes the pointed to item
to appear in the category.

This commit also adds the backend capability to 'pin' content in
a category.  This will be used to support the 'limelight' behaviour
of mygov.