MGS-5844 ensure correct gtm setting pers environment
The values to use for google tag manager are now stored in the cms
as publishing:analytics documents under <site>/administration/analytics.
Each site will have one for dev and one for live.

These documents contain values for container id, auth and env.

Each mount now also contains a property 'publishing:gtm' which is a path
to the right document to use to obtain gtm values.