MGS-6017 Ensure 404 pages can use channel info-based styling
Previously, the EssentialsPageNotFoundComponent was used when serving a 404.
This bypassed the DetermineStylingComponent, and meant that the channel
configuration for styling wasn't available for use. This commit adds a new
PageNotFoundComponent which extends the styling component, and adds 404
behaviour like the essentials component does.

Also removed an unnecessary try/catch block in the styling component.