MGS-6578 prevent title being blank in csv export
previously the title column of the csv export was being populated
with the value of the hippo:name property.  It will now use the
publishing:title property when it is available, and fall back on
hippo:name and then finally  on the name of the node if neither
property is available.