MGS-8033 Ensure cross-site links use correct domain names
Add an hst:types property on mounts to the primary mount for that site.
For live sites, this is the public-facing mount. For non-live sites,
this is the internal-only domain name.

When Bloomreach generates links, it will prefer to link to mounts with
matching types. This change ensures the correct mount is selected for
links, rather than relying on the alphabetical order of domain names.